Moved User state changes to App.tsx
