+++ src/App.tsx
@@ -3,7 +3,7 @@ import "./Indicator.css";
 import "./Canvas.css";
 import "./Header.css";
 import { Route, Routes, useNavigate } from "react-router-dom";
-import { take, timeout } from "rxjs/operators";
+import { map, take, tap, timeout } from "rxjs/operators";
 import { GoalResponse, UserResponse } from "reactive-beeminder-client/dist/api";
 import { Client, IClient } from "reactive-beeminder-client/dist/client";
 import Settings from "./Settings";
@@ -11,6 +11,7 @@ import { getStringKey, storeStringKey } from "./utils/local-storage";
 import Login from "./Login";
 import UserContext from "./contexts/user-context";
 import Year from "./pages/year/Year";
+import { useContext, useEffect } from "react";
 
 const beeminderFetchClient: (t: string) => IClient = (token: string) => ({
   getGoal: (goalName, cb) => {
@@ -62,6 +63,7 @@ export type AppGoal = {
 
 function App() {
   const navigate = useNavigate();
+  const user = useContext(UserContext);
 
   const handleBeeminderTokenChanged = (apiToken: string) => {
     client.setToken(apiToken);
@@ -78,6 +80,20 @@ function App() {
     });
     handleBeeminderTokenChanged(token);
   };
+
+  const getGoalNames = () => {
+    client.userDataStream$
+      .pipe(
+        map(user => user.goals),
+        tap(_ => console.log(_)),
+        take(1)
+      )
+      .subscribe(goalSlugs => user.setGoalSlugs(goalSlugs));
+  };
+
+  useEffect(() => {
+    getGoalNames();
+  });
 
   return (
     <div className="App">
