+++ src/pages/year/Year.tsx
@@ -4,7 +4,6 @@ import { map, take, tap } from "rxjs/operators";
 import { AppGoal } from "../../App";
 import UserContext from "../../contexts/user-context";
 import GoalList from "../../GoalList";
-import { getJsonKey } from "../../utils/local-storage";
 import Canvas from "./components/Canvas";
 
 type YearProps = {
@@ -12,28 +11,10 @@ type YearProps = {
 };
 
 const Year = (props: YearProps) => {
-  const [isAddingGoal, setIsAddingGoal] = useState(true);
-  const [selectedGoals, setSelectedGoals] = useState(readGoalsFromStorage());
+  const [isAddingGoal, setIsAddingGoal] = useState(false);
+  const [selectedGoals, setSelectedGoals] = useState([] as AppGoal[]);
 
   const user = useContext(UserContext);
-
-  const getGoalNames = () => {
-    props.client.userDataStream$
-      .pipe(
-        map(user => user.goals),
-        tap(_ => console.log(_)),
-        take(1)
-      )
-      .subscribe(goalSlugs => {
-        // TODO: Also refer to context here?
-        user.setGoalSlugs(goalSlugs);
-      });
-  };
-
-  useEffect(() => {
-    console.log("asd");
-    getGoalNames();
-  });
 
   return (
     <div>
@@ -53,11 +34,9 @@ const Year = (props: YearProps) => {
           closeClicked={() => setIsAddingGoal(false)}
           goalSelected={slug => {
             setSelectedGoals([...selectedGoals, { slug: slug }]);
-            user.setGoalSlugs([...user.goalSlugs, slug]);
           }}
           goalUnselected={s => {
             setSelectedGoals(selectedGoals.filter(g => g.slug !== s));
-            user.setGoalSlugs(user.goalSlugs.filter(slug => slug !== s));
           }}
           goalSlugs={user.goalSlugs}
         ></GoalList>
@@ -65,10 +44,5 @@ const Year = (props: YearProps) => {
     </div>
   );
 };
+
 export default Year;
-
-function readGoalsFromStorage(): AppGoal[] {
-  const goals: AppGoal[] = getJsonKey("goals");
-
-  return goals ?? [];
-}
